requirements.txt
#flask
Flask[async]

#flask exxtensions
Flask-Cors
flask-restx

# env
python-dotenv

#others
beautifulsoup4
requests


